State store options (#541)
* states option implementation

This add the following options to states:
- persist to local storage
- encrypt
- expire

* add cleanup method

* fix typo
