+++ pkg/app/engine.go
@@ -110,12 +110,16 @@ func (e *engine) Handle(actionName string, src UI, h ActionHandler) {
 	e.actions.handle(actionName, false, src, h)
 }
 
-func (e *engine) SetState(state string, v interface{}) {
-	e.states.Set(state, v)
+func (e *engine) SetState(state string, v interface{}, opts ...StateOption) {
+	e.states.Set(state, v, opts...)
 }
 
 func (e *engine) GetState(state string, recv interface{}) {
 	e.states.Get(state, recv)
+}
+
+func (e *engine) DelState(state string) {
+	e.states.Del(state)
 }
 
 func (e *engine) ObserveState(state string, elem UI) Observer {
@@ -315,8 +319,8 @@ func (e *engine) start(ctx context.Context) {
 		updates := time.NewTicker(currentInterval)
 		defer updates.Stop()
 
-		cleanActions := time.NewTicker(time.Minute)
-		defer cleanActions.Stop()
+		cleanup := time.NewTicker(time.Minute)
+		defer cleanup.Stop()
 
 		for {
 			select {
@@ -345,8 +349,9 @@ func (e *engine) start(ctx context.Context) {
 					updates.Reset(currentInterval)
 				}
 
-			case <-cleanActions.C:
+			case <-cleanup.C:
 				e.Async(e.actions.closeUnusedHandlers)
+				e.Async(e.states.Cleanup)
 			}
 		}
 	})
